+++ src/main/python/fbs_runtime/__init__.py
@@ -1,4 +1,0 @@
-PUBLIC_SETTINGS = {
-	'environment': 'local',
-	'version': '1.7.5'
-}